Creé el componente de la App
